✨ feat: Dropdown StoryBook 추가 및 테스트 코드 제거
